# docs/api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRecord"
                  },
                  "type": "array",
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit",
        "operationId": "submit_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "summary": "Task",
        "operationId": "task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Tasks  Task Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/approval": {
      "post": {
        "summary": "Approval",
        "operationId": "approval_tasks__task_id__approval_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approval Tasks  Task Id  Approval Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/context": {
      "get": {
        "summary": "Context",
        "operationId": "context_tasks__task_id__context_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Context Tasks  Task Id  Context Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/memory/{memory_id}/invalidate": {
      "post": {
        "summary": "Invalidate",
        "operationId": "invalidate_memory__memory_id__invalidate_post",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Memory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Invalidate Memory  Memory Id  Invalidate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Acceptance": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/Check"
            },
            "type": "array",
            "title": "Checks"
          },
          "expected_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Files"
          },
          "contains": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Contains"
          },
          "require_diff": {
            "type": "boolean",
            "title": "Require Diff",
            "default": true
          }
        },
        "type": "object",
        "title": "Acceptance"
      },
      "ApprovalRequest": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "approved"
        ],
        "title": "ApprovalRequest"
      },
      "Budget": {
        "properties": {
          "max_steps": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 24
          },
          "timeout_seconds": {
            "type": "number",
            "maximum": 1800.0,
            "exclusiveMinimum": 0.0,
            "title": "Timeout Seconds",
            "default": 120
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 64.0,
            "title": "Max Tokens",
            "default": 40000
          },
          "max_cost_usd": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cost Usd"
          },
          "max_repairs": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Max Repairs",
            "default": 2
          },
          "provider_retries": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Provider Retries",
            "default": 2
          },
          "context_tokens": {
            "type": "integer",
            "maximum": 32000.0,
            "minimum": 64.0,
            "title": "Context Tokens",
            "default": 6000
          }
        },
        "type": "object",
        "title": "Budget"
      },
      "Check": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "command": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Command"
          },
          "layer": {
            "type": "string",
            "enum": [
              "syntax",
              "unit",
              "integration",
              "repository",
              "acceptance"
            ],
            "title": "Layer",
            "default": "unit"
          }
        },
        "type": "object",
        "required": [
          "name",
          "command"
        ],
        "title": "Check"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "State": {
        "type": "string",
        "enum": [
          "UNDERSTAND",
          "PLAN",
          "EXECUTE",
          "VERIFY",
          "REPAIR",
          "FINISH",
          "FAILED"
        ],
        "title": "State"
      },
      "TaskRecord": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "queued"
          },
          "state": {
            "$ref": "#/components/schemas/State",
            "default": "UNDERSTAND"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "repository_commit": {
            "type": "string",
            "title": "Repository Commit",
            "default": ""
          },
          "workspace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace"
          },
          "request": {
            "$ref": "#/components/schemas/TaskRequest"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count",
            "default": 0
          },
          "repair_count": {
            "type": "integer",
            "title": "Repair Count",
            "default": 0
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "type": "array",
            "title": "Trace"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "working_memory": {
            "additionalProperties": true,
            "type": "object",
            "title": "Working Memory"
          },
          "final_diff": {
            "type": "string",
            "title": "Final Diff",
            "default": ""
          },
          "test_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Result"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "evidence_label": {
            "type": "string",
            "title": "Evidence Label",
            "default": "harness-demonstration"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "repository",
          "request",
          "budget"
        ],
        "title": "TaskRecord"
      },
      "TaskRequest": {
        "properties": {
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "task": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 3,
            "title": "Task"
          },
          "ref": {
            "type": "string",
            "title": "Ref",
            "default": "HEAD"
          },
          "provider": {
            "type": "string",
            "enum": [
              "fake",
              "openai",
              "replay"
            ],
            "title": "Provider",
            "default": "fake"
          },
          "skill": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill",
            "default": "bug-fix"
          },
          "context_strategy": {
            "type": "string",
            "enum": [
              "naive",
              "structured"
            ],
            "title": "Context Strategy",
            "default": "structured"
          },
          "memory_enabled": {
            "type": "boolean",
            "title": "Memory Enabled",
            "default": true
          },
          "transport": {
            "type": "string",
            "enum": [
              "native",
              "mcp"
            ],
            "title": "Transport",
            "default": "native"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "acceptance": {
            "$ref": "#/components/schemas/Acceptance"
          },
          "replay_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replay Id"
          },
          "allow_repository_code": {
            "type": "boolean",
            "title": "Allow Repository Code",
            "default": false
          },
          "use_verifier": {
            "type": "boolean",
            "title": "Use Verifier",
            "default": true
          },
          "use_context": {
            "type": "boolean",
            "title": "Use Context",
            "default": true
          },
          "single_shot": {
            "type": "boolean",
            "title": "Single Shot",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "repository",
          "task"
        ],
        "title": "TaskRequest"
      },
      "ToolResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "output": {
            "type": "string",
            "title": "Output",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "duration_ms": {
            "type": "number",
            "title": "Duration Ms",
            "default": 0
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ToolResult"
      },
      "TraceEvent": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence",
            "default": 0
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "state": {
            "$ref": "#/components/schemas/State"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool"
          },
          "transport": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transport"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_ms": {
            "type": "number",
            "title": "Duration Ms",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "state",
          "kind",
          "summary"
        ],
        "title": "TraceEvent"
      },
      "Usage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          }
        },
        "type": "object",
        "title": "Usage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}